requirements.txt
pyTelegramBotAPI==4.12.0
numpy~=1.24.3
matplotlib~=3.7.1
telebot~=0.0.5
opencv-python~=4.7.0.72
pandas~=1.5.3
ultralytics~=8.0.118
Pillow~=9.5.0
dill~=0.3.6